ws2812: cast clock before dividing for the PIO clkdiv
freq is uint, so 125000000 / (800000 * 10) did integer division and
truncated 15.625 to 15, about 4% out on bit timing. apa102.cpp next door
already casts the numerator.

Fixes #1134

+++ drivers/plasma/ws2812.cpp
@@ -19,7 +19,7 @@ WS2812::WS2812(uint num_leds, PIO pio, uint sm, uint pin, uint freq, bool rgbw, 
     sm_config_set_fifo_join(&c, PIO_FIFO_JOIN_TX);
 
     int cycles_per_bit = ws2812_T1 + ws2812_T2 + ws2812_T3;
-    float div = clock_get_hz(clk_sys) / (freq * cycles_per_bit);
+    float div = (float)clock_get_hz(clk_sys) / (freq * cycles_per_bit);
     sm_config_set_clkdiv(&c, div);
 
     pio_sm_init(pio, sm, pio_program_offset, &c);
